refactor(frontend): decouple business logic from Productivity screen
Extracted state management, filtering, and data orchestration from `frontend/app/(main)/productivity.tsx` into a dedicated ViewModel hook (`useProductivityViewModel`). This refactor addresses a "God Class" architectural defect, adhering strictly to the SOLID implementation directives and the "Logic Leaks" elimination rule.

Debt Identified: The `ProductivityScreen` component violated the Single Responsibility Principle by orchestrating API calls, managing complex local state, handling filtering/sorting, and generating business logic (`generateTodayPlan`) alongside rendering the UI.
Structural Note: Moving state and logic to `useProductivityViewModel` drastically reduces the cognitive load of the screen component, fully decoupling the UI from data orchestration. This allows for easier testing of the view model in isolation and sets up a clean pattern for other large screens.
